Extract room logic into composables
Add three composables (useRoomCRUD, useRoomFilters, useRoomForm) to encapsulate room data loading/deletion, filtering/search logic, and form/modal handling respectively. Refactor app/pages/admin/room_viewer.vue to import and use these composables, removing duplicated types and inline logic and keeping lifecycle mounting to load rooms. useRoomCRUD includes loading/error state and a placeholder deleteRoom message (DELETE not yet implemented on backend); useRoomForm handles validation, modal state, and POST submission with an onRoomSaved callback; useRoomFilters exposes reactive filter state and computed filteredRooms and roomTypeOptions. This separates concerns and simplifies the page component for easier maintenance and reuse.
